Split pi-codex provider helpers
